requirements.txt
openai
anthropic
python-dotenv
chromadb
sentence-transformers
uvicorn
fastapi
pandas
numpy
requests
scikit-learn
torch
torchvision
torchaudio
kobert-transformers
python-multipart
pydantic
pydantic-settings
